Moving to morpheus/testhelpers/sweep package
