+++ morpheus/framework/resources/instance/sweep.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/HewlettPackard/hpe-morpheus-go-sdk/oapigen/sdk"
 
-	"github.com/HPE/terraform-provider-hpe/morpheus/testhelpers"
+	testsweep "github.com/HPE/terraform-provider-hpe/morpheus/testhelpers/sweep"
 )
 
 // Instances whose name begins with this string will be eligible for deletion
@@ -48,7 +48,7 @@ func hasRequiredTags(tags []sdk.AddInstance200ResponseAllOfOneOfInstanceTagsInne
 }
 
 func init() {
-	testhelpers.RegisterTypedAPISweeper(
+	testsweep.RegisterTypedAPISweeper(
 		"hpe_morpheus_instance",
 		// List all instance resources.
 		func(ctx context.Context, client *sdk.APIClient) (
@@ -102,7 +102,7 @@ func init() {
 
 			return &http.Response{StatusCode: http.StatusOK}, nil
 		},
-		testhelpers.WithFilter(func(
+		testsweep.WithFilter(func(
 			ctx context.Context,
 			client *sdk.APIClient,
 			instance sdk.ListInstances200ResponseAllOfInstancesInner,
