+++ morpheus/framework/resources/loadbalancer/sweep.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/HewlettPackard/hpe-morpheus-go-sdk/oapigen/sdk"
 
-	"github.com/HPE/terraform-provider-hpe/morpheus/testhelpers"
+	testsweep "github.com/HPE/terraform-provider-hpe/morpheus/testhelpers/sweep"
 )
 
 // testLoadBalancerPrefix is truncated to 16 characters to account for the
@@ -18,7 +18,7 @@ import (
 const testLoadBalancerPrefix = "TestAccMorpheusL"
 
 func init() {
-	testhelpers.RegisterTypedAPISweeper(
+	testsweep.RegisterTypedAPISweeper(
 		"hpe_morpheus_load_balancer",
 		// List all load balancer resources.
 		func(ctx context.Context, client *sdk.APIClient) (
