+++ morpheus/framework/resources/user/sweep.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/HewlettPackard/hpe-morpheus-go-sdk/oapigen/sdk"
 
-	"github.com/HPE/terraform-provider-hpe/morpheus/testhelpers"
+	testsweep "github.com/HPE/terraform-provider-hpe/morpheus/testhelpers/sweep"
 )
 
 // Users whose name begins with this string will be eligible for deletion
@@ -27,7 +27,7 @@ func isSweepableEmail(email string) bool {
 }
 
 func init() {
-	testhelpers.RegisterTypedAPISweeper(
+	testsweep.RegisterTypedAPISweeper(
 		"hpe_morpheus_user",
 		// List all user resources.
 		func(ctx context.Context, client *sdk.APIClient) (
